[FIX] missing newline in generated files
In a proper file each line end with a newline character, even the last
one.

Wrong file:

Using `write()` function does not add a newline character at the end of
the write. So writing a string in a file with `write()` leads to a file
like this:

```
line1\n
line2\n
lastline
```

Good file:

Using the `print()` command automatically ends the string with a newline
character. So that we ends with a proper file:
```
line1\n
line2\n
lastline\n
```

Also the with statement automatically closes the file at the end of the
block.